backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pymongo>=4.8.0
pydantic>=2.7.4
pydantic[email]>=2.7.4
python-dotenv>=1.0.1
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
python-multipart>=0.0.9
reportlab>=4.2.0
pandas>=2.2.0
openpyxl>=3.1.2
